routes/requirements.txt
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
coverage==7.3.0
exceptiongroup==1.1.2
Flask==2.3.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.7
pytest==7.2.2
pytest-cov==4.0.0
pytest-env==0.8.1
pytest-mock==3.10.0
python-abc==0.2.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.19
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
zipp==3.16.2
sqlalchemy_utils
